# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Digital Marketing Simulation Lab API",
    "description": "API documentation for Digital Marketing Simulation Lab backend",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/admin/billing/plans": {
      "post": {
        "tags": [
          "Admin Billing"
        ],
        "description": "Provisions a new billing/pricing plan (Admin only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "code",
                  "priceMonthly",
                  "priceYearly",
                  "simulationLimit",
                  "studentLimit",
                  "instructorLimit",
                  "certificateLimit",
                  "reportExportLimit"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  },
                  "priceMonthly": {
                    "type": "number"
                  },
                  "priceYearly": {
                    "type": "number"
                  },
                  "currency": {
                    "type": "string",
                    "default": "INR"
                  },
                  "simulationLimit": {
                    "type": "integer"
                  },
                  "studentLimit": {
                    "type": "integer"
                  },
                  "instructorLimit": {
                    "type": "integer"
                  },
                  "certificateLimit": {
                    "type": "integer"
                  },
                  "reportExportLimit": {
                    "type": "integer"
                  },
                  "storageLimitMb": {
                    "type": "integer",
                    "default": 50
                  },
                  "features": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  },
                  "durationDays": {
                    "type": "integer",
                    "default": 30
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/admin/billing/plans/{id}": {
      "put": {
        "tags": [
          "Admin Billing"
        ],
        "description": "Updates details of an existing billing plan (Admin only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "priceMonthly": {
                    "type": "number"
                  },
                  "priceYearly": {
                    "type": "number"
                  },
                  "currency": {
                    "type": "string"
                  },
                  "simulationLimit": {
                    "type": "integer"
                  },
                  "studentLimit": {
                    "type": "integer"
                  },
                  "instructorLimit": {
                    "type": "integer"
                  },
                  "certificateLimit": {
                    "type": "integer"
                  },
                  "reportExportLimit": {
                    "type": "integer"
                  },
                  "storageLimitMb": {
                    "type": "integer"
                  },
                  "features": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "durationDays": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/admin/billing/subscriptions": {
      "get": {
        "tags": [
          "Admin Billing"
        ],
        "description": "Retrieve subscription reports with status and search filters (Admin only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "active",
                "expiring",
                "cancelled",
                "trial"
              ],
              "default": "all"
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "subscriptions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "planId": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "billingCycle": {
                            "type": "string"
                          },
                          "startDate": {
                            "type": "string"
                          },
                          "endDate": {
                            "type": "string"
                          },
                          "cancelAtPeriodEnd": {
                            "type": "boolean"
                          },
                          "gatewaySubscriptionId": {
                            "type": "string",
                            "nullable": true
                          },
                          "gatewayCustomerId": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "user": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              },
                              "role": {
                                "type": "string"
                              }
                            }
                          },
                          "plan": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              },
                              "priceMonthly": {
                                "type": "number"
                              },
                              "priceYearly": {
                                "type": "number"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ]
}